Add data for 2020-08-12, not including MA county data
Since Massachusetts no longer reports per-county data on a daily basis, and has changed how it reports both cases and
deaths even in the weekly report, it's not possible to continue including per-county data.  Similarly, probable cases
are no longer included in the daily reports, although probable deaths still are.  Usage of total and daily cases will
need to be modified to include only confirmed cases to avoid inconsistent data, and per-county numbers will no longer
be updated as of 2020-08-12.